requirements.txt
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
attrs==25.3.0
autodistill==0.1.29
autodistill-detr==0.1.1
autodistill-grounding-dino==0.1.4
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
filetype==1.2.0
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.2
huggingface-hub==0.30.2
idna==3.7
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
lightning-utilities==0.14.3
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
multidict==6.4.3
networkx==3.4.2
numpy==2.2.5
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
packaging==25.0
pillow==11.2.1
pillow_heif==0.22.0
platformdirs==4.3.7
propcache==0.3.1
pycocotools==2.0.8
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytorch-lightning==2.5.1.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rf-segment-anything==1.0
rf_groundingdino==0.1.2
roboflow==1.1.61
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
setuptools==80.0.0
six==1.17.0
supervision==0.9.0
sympy==1.13.1
threadpoolctl==3.6.0
timm==1.0.15
tokenizers==0.21.1
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchmetrics==1.7.1
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.51.3
typing_extensions==4.13.2
urllib3==2.4.0
yapf==0.43.0
yarl==1.20.0
